+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
-- New features under development.
+- Official Pump Fun parser (https://github.com/rpcpool/yellowstone-vixen/pull/66)
+- Subscribe to block_meta (https://github.com/rpcpool/yellowstone-vixen/pull/67)
 
 ### Changed